Initial Version of Runes Keyboard
This is my first keyboard. Please forgive me if I have missed something; I'm still learning this process. The purpose of this keyboard is to allow the user to type with runes, and there are many deadkeys which allow for the selection of different runes from different runic alphabets such as Anglo-Saxon and Old Norse.